split: extract ApplicationManagementService from AdminService
Moves updateApplication, setApplicationEnabled, regenerateClientSecret to
a focused service. AdminService drops applicationRepository, passwordHasher,
and corsPort dependencies.
