+++ src/main/kotlin/com/kauth/Application.kt
@@ -446,6 +446,7 @@ fun Application.module(
             roleGroupService = s.roleGroupService,
             adminService = s.adminService,
             adminUserService = s.adminUserService,
+            applicationManagementService = s.applicationManagementService,
             userAttributeService = s.userAttributeService,
             claimMapperService = s.claimMapperService,
             emailOtpService = s.emailOtpService,
@@ -469,6 +470,7 @@ fun Application.module(
             adminService = s.adminService,
             workspaceSettingsService = s.workspaceSettingsService,
             adminUserService = s.adminUserService,
+            applicationManagementService = s.applicationManagementService,
             roleGroupService = s.roleGroupService,
             appInfo = appInfo,
             tenantRepository = s.tenantRepository,
