+++ src/main/kotlin/com/kauth/domain/service/AdminService.kt
@@ -1,259 +1,23 @@
 package com.kauth.domain.service
 
-import com.kauth.domain.model.Application
-import com.kauth.domain.model.ApplicationId
 import com.kauth.domain.model.AuditEvent
 import com.kauth.domain.model.AuditEventType
 import com.kauth.domain.model.RequiredAction
 import com.kauth.domain.model.Tenant
 import com.kauth.domain.model.TenantId
 import com.kauth.domain.model.UserId
-import com.kauth.domain.port.ApplicationRepository
 import com.kauth.domain.port.AuditLogPort
-import com.kauth.domain.port.CorsPort
-import com.kauth.domain.port.PasswordHasher
 import com.kauth.domain.port.PasswordPolicyPort
 import com.kauth.domain.port.TenantRepository
 import com.kauth.domain.port.UserRepository
-import com.kauth.domain.util.SecureTokens
 
 class AdminService(
     private val tenantRepository: TenantRepository,
     private val userRepository: UserRepository,
-    private val applicationRepository: ApplicationRepository,
-    private val passwordHasher: PasswordHasher,
     private val auditLog: AuditLogPort,
     private val selfServiceService: UserSelfServiceService,
     private val passwordPolicy: PasswordPolicyPort? = null,
-    private val corsPort: CorsPort? = null,
 ) {
-    private fun invalidateCors(tenantId: TenantId) {
-        val port = corsPort ?: return
-        tenantRepository.findById(tenantId)?.slug?.let(port::invalidate)
-    }
-
-    /**
-     * Returns the http(s) origin (scheme://host[:port]) of [url], or null if the
-     * URL is not a parseable absolute http/https URL. Used to compare launcher
-     * URL origins against registered redirect URI origins.
-     */
-    private fun extractHttpOrigin(url: String): String? =
-        try {
-            val uri = java.net.URI(url)
-            val scheme = uri.scheme?.lowercase()
-            val host = uri.host
-            if (scheme != "http" && scheme != "https") {
-                null
-            } else if (host.isNullOrBlank()) {
-                null
-            } else {
-                val port = if (uri.port != -1) ":${uri.port}" else ""
-                "$scheme://$host$port"
-            }
-        } catch (_: Exception) {
-            null
-        }
-
-    fun updateApplication(
-        appId: ApplicationId,
-        tenantId: TenantId,
-        name: String? = null,
-        description: String? = null,
-        accessType: String? = null,
-        redirectUris: List<String>? = null,
-        launcherUrl: String? = null,
-        iconUrl: String? = null,
-        launcherVisible: Boolean? = null,
-        launcherDisplayOrder: Int? = null,
-        audience: String? = null,
-    ): AdminResult<Application> {
-        val app =
-            applicationRepository.findById(appId)
-                ?: return AdminResult.Failure(AdminError.NotFound("Application not found."))
-
-        if (app.tenantId != tenantId) {
-            return AdminResult.Failure(AdminError.NotFound("Application not found in this workspace."))
-        }
-
-        val resolvedName = name?.trim() ?: app.name
-        val resolvedDescription =
-            if (description !=
-                null
-            ) {
-                description.trim().takeIf { it.isNotBlank() }
-            } else {
-                app.description
-            }
-        val resolvedAccessType = accessType ?: app.accessType.value
-        val resolvedRedirectUris = redirectUris ?: app.redirectUris
-        val resolvedLauncherUrl =
-            if (launcherUrl != null) launcherUrl.trim().takeIf { it.isNotBlank() } else app.launcherUrl
-        val resolvedIconUrl =
-            if (iconUrl != null) iconUrl.trim().takeIf { it.isNotBlank() } else app.iconUrl
-        val resolvedLauncherVisible = launcherVisible ?: app.launcherVisible
-        val resolvedLauncherDisplayOrder = launcherDisplayOrder ?: app.launcherDisplayOrder
-        val resolvedAudience =
-            if (audience != null) audience.trim().takeIf { it.isNotBlank() } else app.audience
-
-        if (resolvedName.isBlank()) {
-            return AdminResult.Failure(AdminError.Validation("Name is required."))
-        }
-
-        if (resolvedRedirectUris.isEmpty()) {
-            return AdminResult.Failure(
-                AdminError.Validation(
-                    "At least one redirect URI is required. The authorization code flow " +
-                        "(including email-OTP back-channel exchange) needs a registered URI to bind to.",
-                ),
-            )
-        }
-
-        if (resolvedAudience != null && resolvedAudience.length > 200) {
-            return AdminResult.Failure(AdminError.Validation("Token audience must be 200 characters or fewer."))
-        }
-
-        // Origin validation prevents a phishing surface where a compromised admin sets the
-        // tile to an attacker-controlled host. Mitigation: launcher URL must share origin
-        // with one of the application's already-registered redirect URIs.
-        if (resolvedLauncherUrl != null) {
-            val launcherOrigin =
-                extractHttpOrigin(resolvedLauncherUrl)
-                    ?: return AdminResult.Failure(
-                        AdminError.Validation("Launcher URL must be a valid http or https URL."),
-                    )
-            val redirectOrigins = resolvedRedirectUris.mapNotNull { extractHttpOrigin(it) }.toSet()
-            if (redirectOrigins.isEmpty()) {
-                return AdminResult.Failure(
-                    AdminError.Validation(
-                        "Add at least one redirect URI on the same origin before setting a launcher URL.",
-                    ),
-                )
-            }
-            if (launcherOrigin !in redirectOrigins) {
-                return AdminResult.Failure(
-                    AdminError.Validation(
-                        "Launcher URL origin '$launcherOrigin' must match one of the registered redirect URI origins.",
-                    ),
-                )
-            }
-        }
-
-        if (resolvedIconUrl != null && extractHttpOrigin(resolvedIconUrl) == null) {
-            return AdminResult.Failure(AdminError.Validation("Icon URL must be a valid http or https URL."))
-        }
-
-        val updated =
-            applicationRepository.update(
-                appId = appId,
-                name = resolvedName,
-                description = resolvedDescription,
-                accessType = resolvedAccessType,
-                redirectUris = resolvedRedirectUris,
-                launcherUrl = resolvedLauncherUrl,
-                iconUrl = resolvedIconUrl,
-                launcherVisible = resolvedLauncherVisible,
-                launcherDisplayOrder = resolvedLauncherDisplayOrder,
-                audience = resolvedAudience,
-            )
-
-        invalidateCors(tenantId)
-
-        auditLog.record(
-            AuditEvent(
-                tenantId = tenantId,
-                userId = null,
-                clientId = appId,
-                eventType = AuditEventType.ADMIN_CLIENT_UPDATED,
-                ipAddress = null,
-                userAgent = null,
-                details = mapOf("clientId" to app.clientId),
-            ),
-        )
-
-        return AdminResult.Success(updated)
-    }
-
-    /**
-     * Enables or disables an application.
-     */
-    fun setApplicationEnabled(
-        appId: ApplicationId,
-        tenantId: TenantId,
-        enabled: Boolean,
-    ): AdminResult<Unit> {
-        val app =
-            applicationRepository.findById(appId)
-                ?: return AdminResult.Failure(AdminError.NotFound("Application not found."))
-
-        if (app.tenantId != tenantId) {
-            return AdminResult.Failure(AdminError.NotFound("Application not found in this workspace."))
-        }
-
-        applicationRepository.setEnabled(appId, enabled)
-
-        invalidateCors(tenantId)
-
-        auditLog.record(
-            AuditEvent(
-                tenantId = tenantId,
-                userId = null,
-                clientId = appId,
-                eventType = if (enabled) AuditEventType.ADMIN_CLIENT_ENABLED else AuditEventType.ADMIN_CLIENT_DISABLED,
-                ipAddress = null,
-                userAgent = null,
-                details = mapOf("clientId" to app.clientId, "enabled" to enabled.toString()),
-            ),
-        )
-
-        return AdminResult.Success(Unit)
-    }
-
-    /**
-     * Generates a new client secret for a confidential application.
-     * Returns the raw secret — it will NOT be shown again.
-     * Only the bcrypt hash is persisted.
-     */
-    fun regenerateClientSecret(
-        appId: ApplicationId,
-        tenantId: TenantId,
-    ): AdminResult<String> {
-        val app =
-            applicationRepository.findById(appId)
-                ?: return AdminResult.Failure(AdminError.NotFound("Application not found."))
-
-        if (app.tenantId != tenantId) {
-            return AdminResult.Failure(AdminError.NotFound("Application not found in this workspace."))
-        }
-
-        // Generate 32-byte (256-bit) cryptographically random secret, base64url encoded
-        val secret = SecureTokens.randomBase64Url(32)
-
-        applicationRepository.setClientSecretHash(appId, passwordHasher.hash(secret))
-
-        auditLog.record(
-            AuditEvent(
-                tenantId = tenantId,
-                userId = null,
-                clientId = appId,
-                eventType = AuditEventType.ADMIN_CLIENT_SECRET_REGENERATED,
-                ipAddress = null,
-                userAgent = null,
-                details = mapOf("clientId" to app.clientId),
-            ),
-        )
-
-        return AdminResult.Success(secret)
-    }
-
-    // =========================================================================
-    // SMTP configuration
-    // =========================================================================
-
-    /**
-     * Updates SMTP configuration for a workspace.
-     * [smtpPassword] should be the raw (plaintext) value — the adapter layer
-     * encrypts it before persistence via the infrastructure EncryptionService.
-     */
     fun updateSmtpConfig(
         slug: String,
         smtpHost: String?,
