+++ src/test/kotlin/com/kauth/adapter/web/admin/AdminApiKeysTest.kt
@@ -124,8 +124,6 @@ class AdminApiKeysTest {
         AdminService(
             tenantRepository = tenantRepo,
             userRepository = userRepo,
-            applicationRepository = appRepo,
-            passwordHasher = hasher,
             auditLog = auditLogPort,
             selfServiceService = buildSelfService(),
         )
@@ -138,6 +136,14 @@ class AdminApiKeysTest {
             passwordHasher = hasher,
             auditLog = auditLogPort,
             selfServiceService = buildSelfService(),
+        )
+
+    private fun buildAppMgmtService() =
+        com.kauth.domain.service.ApplicationManagementService(
+            applicationRepository = appRepo,
+            tenantRepository = tenantRepo,
+            passwordHasher = hasher,
+            auditLog = auditLogPort,
         )
 
     private fun buildRoleGroupService() =
@@ -297,6 +303,7 @@ class AdminApiKeysTest {
                     com.kauth.domain.service
                         .WorkspaceSettingsService(tenantRepo, auditLogPort),
                 adminUserService = buildAdminUserService(),
+                applicationManagementService = buildAppMgmtService(),
                 roleGroupService = buildRoleGroupService(),
                 appInfo = AppInfo(),
                 tenantRepository = tenantRepo,
