+++ src/test/kotlin/com/kauth/adapter/web/admin/AdminRoutesTest.kt
@@ -110,8 +110,6 @@ class AdminRoutesTest {
         AdminService(
             tenantRepository = tenantRepo,
             userRepository = userRepo,
-            applicationRepository = appRepo,
-            passwordHasher = hasher,
             auditLog = auditLogPort,
             selfServiceService = buildSelfService(),
         )
@@ -124,6 +122,14 @@ class AdminRoutesTest {
             passwordHasher = hasher,
             auditLog = auditLogPort,
             selfServiceService = buildSelfService(),
+        )
+
+    private fun buildAppMgmtService() =
+        com.kauth.domain.service.ApplicationManagementService(
+            applicationRepository = appRepo,
+            tenantRepository = tenantRepo,
+            passwordHasher = hasher,
+            auditLog = auditLogPort,
         )
 
     private fun buildRoleGroupService() =
@@ -620,6 +626,7 @@ class AdminRoutesTest {
                     com.kauth.domain.service
                         .WorkspaceSettingsService(tenantRepo, auditLogPort),
                 adminUserService = buildAdminUserService(),
+                applicationManagementService = buildAppMgmtService(),
                 roleGroupService = buildRoleGroupService(),
                 appInfo = AppInfo(),
                 tenantRepository = tenantRepo,
@@ -705,6 +712,7 @@ class AdminRoutesTest {
                     com.kauth.domain.service
                         .WorkspaceSettingsService(tenantRepo, auditLogPort),
                 adminUserService = buildAdminUserService(),
+                applicationManagementService = buildAppMgmtService(),
                 roleGroupService = buildRoleGroupService(),
                 appInfo = AppInfo(),
                 tenantRepository = tenantRepo,
