+++ src/test/kotlin/com/kauth/adapter/web/admin/AdminUserAttributesAndClaimMappersTest.kt
@@ -357,8 +357,6 @@ class AdminUserAttributesAndClaimMappersTest {
         AdminService(
             tenantRepository = tenantRepo,
             userRepository = userRepo,
-            applicationRepository = appRepo,
-            passwordHasher = hasher,
             auditLog = auditLogPort,
             selfServiceService = buildSelfService(),
         )
@@ -371,6 +369,14 @@ class AdminUserAttributesAndClaimMappersTest {
             passwordHasher = hasher,
             auditLog = auditLogPort,
             selfServiceService = buildSelfService(),
+        )
+
+    private fun buildAppMgmtService() =
+        com.kauth.domain.service.ApplicationManagementService(
+            applicationRepository = appRepo,
+            tenantRepository = tenantRepo,
+            passwordHasher = hasher,
+            auditLog = auditLogPort,
         )
 
     private fun buildRoleGroupService() =
@@ -417,6 +423,7 @@ class AdminUserAttributesAndClaimMappersTest {
                     com.kauth.domain.service
                         .WorkspaceSettingsService(tenantRepo, auditLogPort),
                 adminUserService = buildAdminUserService(),
+                applicationManagementService = buildAppMgmtService(),
                 roleGroupService = buildRoleGroupService(),
                 appInfo = AppInfo(),
                 tenantRepository = tenantRepo,
