+++ src/test/kotlin/com/kauth/adapter/web/api/ApiClaimMapperRoutesTest.kt
@@ -348,8 +348,6 @@ class ApiClaimMapperRoutesTest {
                     com.kauth.domain.service.AdminService(
                         tenantRepository = tenantRepo,
                         userRepository = userRepo,
-                        applicationRepository = com.kauth.fakes.FakeApplicationRepository(),
-                        passwordHasher = hasher,
                         auditLog = com.kauth.fakes.FakeAuditLogPort(),
                         selfServiceService = buildFakeSelfService(),
                     ),
@@ -361,6 +359,13 @@ class ApiClaimMapperRoutesTest {
                         passwordHasher = hasher,
                         auditLog = com.kauth.fakes.FakeAuditLogPort(),
                         selfServiceService = buildFakeSelfService(),
+                    ),
+                applicationManagementService =
+                    com.kauth.domain.service.ApplicationManagementService(
+                        applicationRepository = com.kauth.fakes.FakeApplicationRepository(),
+                        tenantRepository = tenantRepo,
+                        passwordHasher = hasher,
+                        auditLog = com.kauth.fakes.FakeAuditLogPort(),
                     ),
                 userAttributeService = userAttributeService,
                 claimMapperService = claimMapperService,
