+++ src/test/kotlin/com/kauth/adapter/web/api/ApiRoutesTest.kt
@@ -137,8 +137,6 @@ class ApiRoutesTest {
         AdminService(
             tenantRepository = tenantRepo,
             userRepository = userRepo,
-            applicationRepository = appRepo,
-            passwordHasher = hasher,
             auditLog = auditLogPort,
             selfServiceService = selfServiceService,
         )
@@ -151,6 +149,14 @@ class ApiRoutesTest {
             passwordHasher = hasher,
             auditLog = auditLogPort,
             selfServiceService = selfServiceService,
+        )
+
+    private val applicationManagementService =
+        com.kauth.domain.service.ApplicationManagementService(
+            applicationRepository = appRepo,
+            tenantRepository = tenantRepo,
+            passwordHasher = hasher,
+            auditLog = auditLogPort,
         )
 
     private val roleGroupService =
@@ -931,6 +937,7 @@ class ApiRoutesTest {
                 roleGroupService = roleGroupService,
                 adminService = adminService,
                 adminUserService = adminUserService,
+                applicationManagementService = applicationManagementService,
                 userAttributeService = userAttributeService,
                 claimMapperService = claimMapperService,
                 emailOtpService = stubEmailOtpService(),
